Environment-sensitive MPI for hood.math
This commit amends the 'hood' system-sepcific makefile to use the MPI
libraries loaded from the user's shell profile.  This is handy because
hood.math supports multiple versions of OpenMPI, which need to be used
consistently.

Now, the makefile generation process on hood locks in the MPI version
**when the makefile is generated**.  This should carry through in using
the makefile-specified MPI version at compile time, but if the user's
environment is inconsistent they'll encounter a runtime error.