Move BC-parsing helper to BaseCase
This commit moves the helper function that process
expansion/boundary conditions ("FOURIER" -> PERIODIC, etc)
to BaseCase.cpp, from Options.

This change should be invisible to most case files, which
habitually include BaseCase for helper functions as-is, but
it allows Options to once again be independent of
NSIntegrator.  This improves modularity for other uses of
the code, such as a hypothetical SPINSoff (get it?  it's a
pun) that would like to do some spectral voodoo with an
option file but doesn't need the full Navier-Stokes
machinery.