Move time step checker to BaseCase
The empty shell of check_timestep in BaseCase has been filled
with what was essentially the same thing in each case file.
Again, this is just more case file house cleaning.

A few ancillary functions have also been written to help.
The most important might be get_dt_max() which returns dt_max.
dt_max can either be specified in spins.conf or defined in the
case file based on the buoyancy frequency. Currently, if
dt_max > 0 and in spins.conf, then that value is used. Otherwise,
0.5/sqrt(N2_max) is used.

The re-definition of check_timestep in each case file can now be
removed so long as the associated ancillary functions and
parameters are included.