Update gravity_current to only use full zgrid
The full xgrid and ygrid are unnecessary since x and y only
depend on a single variable (or index - ii for x, and jj for y).
The variables xx and yy can be used instead of xgrid and ygrid,
respectively. Creating the full grids for these fields is a
waste of memory.