Increase the needed dump write time to account for clock time step
Some simulations have a long clock time per step, but have a short
write to disk time. If only the short write to disk time is
accounted for in the check_and_dump command then insufficient
time may be allocated for the dump to disk. Both are now accounted
for.