Update map_iwave case to use new diagnostics
This case has uncovered that the BPE calculation returns
incorrect values when a Chebyshev grid is used on a field scale.
The BPE calculation uses the quadrature weights as the volumes
of each cell. This is true when the grid is regular, but incorrect
for Chebyshev grids.

To first approximation the Cheb quadrature weights are close to
the volumes of the cells. This leads to the BPE calculation
being accurate to 8 significant figures. 8 sig figs is not enough
on fields scales since the magnitude of the PE can be very, very
large. The difference of the BPE from the PE can thus be smaller
than these 8 sig figs.