Adds version numbering to the repository
The file VERSION now records the MAJOR_VERSION, MINOR_VERSION,
and PATCH_VERSION.
To update the version number, the user only needs to modify the
appropriate lines in VERSION.

Makefile has been updated to automatically pass the version
information through compiler flags.

BaseCase.cpp has been updated to add a print statement into the
initialization function. By default, every casefile that uses
BaseCase will now print the SPINS version with which is was compiled.
This should help avoid issues with mis-matched versions.

It will appear in the output as:
SPINS Version 1.0.1

The starting version is set to 2.0.0

StackOverflow has a nice summary of when to change which version number
(https://stackoverflow.com/questions/3826580/what-rules-does-software-version-numbering-follow)
 - Major version numbers change whenever there is some significant change
being introduced. For example, a large or potentially
backward-incompatible change to a software package.
 - Minor version numbers change when a new, minor feature is introduced or
when a set of smaller features is rolled out.
 - Patch numbers change when a new build of the software is released to
customers. This is normally for small bug-fixes or the like.