Fix topography error in wave_reader
The quadrature weights were found to be incorrect for mapped grids
because they were adjusted using the topography which was not
scaled properly. The topography needed to be adjusted to be relative
of the 'minimum z-value'. This is in quation marks because it is
not necessarily the absolute minimum, but rather the reference
minimum.
This shift also fixed the BPE calculation to have the same reference
as the PE calculation.