Updates to solver parameters
Based on experience with a poorly-converging, steep hill case, the
number of inner iterations in GMRES is increased to 40, and the number
of outer iterations drops to 2.

Please pay attention to GMRES warnings and report cases where things
don't work well, there's probably an optimal parameter range (or else we
could make the parameters a runtime config option).