wave_reader: allow FULL input types
Previously, wave_reader locked 'FULL' (3D) input behind enabling
restart.  This restriction was unnecessary, and doing so in fact made
the specification of u/v/w/rho filenames useless.

Deleting this check allows proper read-in of FULL datatypes, without
faking a restart.  The only functional difference is that if
perturbations are also enabled, they will be applied on top of the 3D
data read.  This seems to be desirable, since it makes the application
of the config-file settings more straightforward.